Fixed a tricky bug to do with the order of looping through parent ops.
The loop through parent_ops in Node.__call__ wasn't supposed to require a
particular ordering. However, if the gradmaker closure used its the inputs
before creating the inner function (e.g. make_grad_eigh), then the number of
tapes associated with the nodes stored in the closure depended on the
ordering. This led to undefined behavior, since the ordering ultimately depended
on the ordering of the keys in the Node.tape dicts.