Removed type checking on gradient evaluations.
This was really nice to have, and it caught a few bugs, but the issue is that
numpy often converts 0-dim arrays to scalars (e.g. `np.sin(np.array(0.0))` gives
a scalar). This throws off type checking, and the fix would be painful. We'd
have to wrap all gradients to check the input type and cast the gradient
accordingly if the type was changed.